added components for header, searchbar and results
